components/update-agent/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "fabric-update-agent"
description = "SingleAxis Fabric — pre-apply manifest verifier (Layer 1)."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.12"
authors = [{ name = "AI5Labs Research OPC Private Limited", email = "fabric@singleaxis.ai" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3.12",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
  "cryptography>=42.0,<44.0",     # Ed25519 signature verification
  "fastapi>=0.112,<1.0",          # admission webhook HTTP server
  "uvicorn>=0.30,<1.0",
  "jsonschema>=4.22,<5.0",
  "packaging>=23.2",              # version constraint parsing
  "pydantic>=2.8,<3.0",
  "pyyaml>=6.0,<7.0",
  "typer>=0.12,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "httpx>=0.27,<1.0",             # FastAPI TestClient
  "ruff>=0.6.9",
  "mypy>=1.11.2",
  "types-PyYAML>=6.0",
  "types-jsonschema>=4.22",
]

[project.scripts]
fabric-update-agent = "fabric_update_agent.__main__:main"

[tool.hatch.version]
path = "src/fabric_update_agent/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/fabric_update_agent"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
  "-ra",
  "--strict-config",
  "--strict-markers",
  "--cov=fabric_update_agent",
  "--cov-report=term-missing",
  "--cov-fail-under=80",
]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["fabric_update_agent"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "S", "C4", "PIE", "RET", "SIM", "PL", "RUF"]
# PLR0911: verifier.verify returns once per check — splitting it into
# helpers would actually make the deny reasons harder to trace.
ignore = ["PLR0913", "PLR0911"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "PLR2004", "S106"]
# Typer's pattern: the function signature's defaults are `typer.Option(...)`.
"src/fabric_update_agent/__main__.py" = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
